requirements.txt
allure-pytest
pytest
pytest-html
pytest-xdist
Appium-Python-Client
appium
webdriver-manager